Created TimerControls Component
