+++ src/pomodoro/Pomodoro.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import classNames from "../utils/class-names";
 import useInterval from "../utils/useInterval";
 import DurationSetters from "./DurationSetters";
+import TimerControls from "./TimerControls";
 import Session from "./Session";
 
 // These functions are defined outside of the component to insure they do not have access to state
@@ -91,36 +91,7 @@ function Pomodoro() {
     isTimerRunning ? 1000 : null
   );
 
-  /**
-   * Called whenever the play/pause button is clicked.
-   */
-  function playPause() {
-    setIsTimerRunning((prevState) => {
-      const nextState = !prevState;
-      if (nextState) {
-        setSession((prevStateSession) => {
-          // If the timer is starting and the previous session is null,
-          // start a focusing session.
-          if (prevStateSession === null) {
-            return {
-              label: "Focusing",
-              timeRemaining: focusObj.duration * 60,
-            };
-          }
-          return prevStateSession;
-        });
-      }
-      return nextState;
-    });
-  }
-
-  function stop() {
-    setIsTimerRunning(() => {
-      setSession(null);
-      return false;
-    });
-  }
-
+  //JSX Return all Components
   return (
     <div className="pomodoro">
       <DurationSetters
@@ -130,41 +101,13 @@ function Pomodoro() {
         breakObj={breakObj}
         setBreakObj={setBreakObj}
       />
-      <div className="row">
-        <div className="col">
-          <div
-            className="btn-group btn-group-lg mb-2"
-            role="group"
-            aria-label="Timer controls"
-          >
-            <button
-              type="button"
-              className="btn btn-primary"
-              data-testid="play-pause"
-              title="Start or pause timer"
-              onClick={playPause}
-            >
-              <span
-                className={classNames({
-                  oi: true,
-                  "oi-media-play": !isTimerRunning,
-                  "oi-media-pause": isTimerRunning,
-                })}
-              />
-            </button>
-            <button
-              type="button"
-              className="btn btn-secondary"
-              data-testid="stop"
-              title="Stop the session"
-              onClick={stop}
-              disabled={!session}
-            >
-              <span className="oi oi-media-stop" />
-            </button>
-          </div>
-        </div>
-      </div>
+      <TimerControls
+        session={session}
+        setSession={setSession}
+        isTimerRunning={isTimerRunning}
+        setIsTimerRunning={setIsTimerRunning}
+        timeRemaining={focusObj.duration * 60}
+      />
       <Session
         session={session}
         focusDuration={focusObj.duration}
